+++ src/interfaces/applicative.ts
@@ -1,5 +1,5 @@
 import { Functor } from './functor';
 
-export interface Applicative<T> extends Functor<T> {
-  apply(other: Applicative<(value: T) => any>): Applicative<any>;
+export interface Applicative extends Functor {
+  apply(other: Applicative): Applicative;
 }